# OpenAPI.yml
...
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: http://localhost:8081
    description: Generated server url
paths:
...
  "/api/means/{sectionId}/{studentId}":
    put:
      tags:
        - mean-controller
      summary: Update a mean by studentId
      operationId: updateMean
      parameters:
        - name: sectionId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: studentId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/Mean"
        required: true
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/Mean"
        "404":
          description: Not Found
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorMessage"
        "500":
          description: Internal Server Error
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorMessage"
...
  "/api/sections/{sectionId}/students/{studentId}/means":
    get:
      tags:
        - mean-controller
      summary: Get a mean by sectionsId and studentId
      operationId: getAllMeansBystudentIdAndsectionId
      parameters:
        - name: studentId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: sectionId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/Mean"
        "404":
          description: Not Found
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorMessage"
        "500":
          description: Internal Server Error
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorMessage"
    post:
      tags:
        - mean-controller
      summary: Create a mean by sectionId and studentId
      operationId: createMean
      parameters:
        - name: sectionId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: studentId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/Mean"
        required: true
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/Mean"
        "404":
          description: Not Found
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorMessage"
        "500":
          description: Internal Server Error
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorMessage"
...
  "/api/students/{studentId}/means":
    get:
      tags:
        - mean-controller
      summary: Get a mean by studentId
      operationId: getAllMeansBystudentId
      parameters:
        - name: studentId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/Mean"
        "404":
          description: Not Found
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorMessage"
        "500":
          description: Internal Server Error
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorMessage"
...
  "/api/sections/{sectionId}/means":
    get:
      tags:
        - mean-controller
      summary: Get a mean by sectionId
      operationId: getAllMeansBysectionId
      parameters:
        - name: sectionId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/Mean"
        "404":
          description: Not Found
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorMessage"
        "500":
          description: Internal Server Error
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorMessage"
...
  "/api/means":
    get:
      tags:
        - mean-controller
      summary: Get all means
      operationId: getAllMeans
      parameters:
        - name: mean
          in: query
          required: false
          schema:
            type: integer
            format: int32
        - name: student
          in: query
          required: false
          schema:
            "$ref": "#/components/schemas/Student"
        - name: section
          in: query
          required: false
          schema:
            "$ref": "#/components/schemas/Section"
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/Mean"
        "404":
          description: Not Found
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorMessage"
        "500":
          description: Internal Server Error
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorMessage"
  "/api/means/{id}":
    get:
      tags:
        - mean-controller
      summary: Get a mean by its id
      operationId: getMeansById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/Mean"
        "404":
          description: Not Found
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorMessage"
        "500":
          description: Internal Server Error
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorMessage"
    delete:
      tags:
        - mean-controller
      summary: Delete a mean by id
      operationId: deleteMean
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                type: string
                enum:
                  - 100 CONTINUE
                  - 101 SWITCHING_PROTOCOLS
                  - 102 PROCESSING
                  - 103 CHECKPOINT
                  - 200 OK
                  - 201 CREATED
                  - 202 ACCEPTED
                  - 203 NON_AUTHORITATIVE_INFORMATION
                  - 204 NO_CONTENT
                  - 205 RESET_CONTENT
                  - 206 PARTIAL_CONTENT
                  - 207 MULTI_STATUS
                  - 208 ALREADY_REPORTED
                  - 226 IM_USED
                  - 300 MULTIPLE_CHOICES
                  - 301 MOVED_PERMANENTLY
                  - 302 FOUND
                  - 302 MOVED_TEMPORARILY
                  - 303 SEE_OTHER
                  - 304 NOT_MODIFIED
                  - 305 USE_PROXY
                  - 307 TEMPORARY_REDIRECT
                  - 308 PERMANENT_REDIRECT
                  - 400 BAD_REQUEST
                  - 401 UNAUTHORIZED
                  - 402 PAYMENT_REQUIRED
                  - 403 FORBIDDEN
                  - 404 NOT_FOUND
                  - 405 METHOD_NOT_ALLOWED
                  - 406 NOT_ACCEPTABLE
                  - 407 PROXY_AUTHENTICATION_REQUIRED
                  - 408 REQUEST_TIMEOUT
                  - 409 CONFLICT
                  - 410 GONE
                  - 411 LENGTH_REQUIRED
                  - 412 PRECONDITION_FAILED
                  - 413 PAYLOAD_TOO_LARGE
                  - 413 REQUEST_ENTITY_TOO_LARGE
                  - 414 URI_TOO_LONG
                  - 414 REQUEST_URI_TOO_LONG
                  - 415 UNSUPPORTED_MEDIA_TYPE
                  - 416 REQUESTED_RANGE_NOT_SATISFIABLE
                  - 417 EXPECTATION_FAILED
                  - 418 I_AM_A_TEAPOT
                  - 419 INSUFFICIENT_SPACE_ON_RESOURCE
                  - 420 METHOD_FAILURE
                  - 421 DESTINATION_LOCKED
                  - 422 UNPROCESSABLE_ENTITY
                  - 423 LOCKED
                  - 424 FAILED_DEPENDENCY
                  - 425 TOO_EARLY
                  - 426 UPGRADE_REQUIRED
                  - 428 PRECONDITION_REQUIRED
                  - 429 TOO_MANY_REQUESTS
                  - 431 REQUEST_HEADER_FIELDS_TOO_LARGE
                  - 451 UNAVAILABLE_FOR_LEGAL_REASONS
                  - 500 INTERNAL_SERVER_ERROR
                  - 501 NOT_IMPLEMENTED
                  - 502 BAD_GATEWAY
                  - 503 SERVICE_UNAVAILABLE
                  - 504 GATEWAY_TIMEOUT
                  - 505 HTTP_VERSION_NOT_SUPPORTED
                  - 506 VARIANT_ALSO_NEGOTIATES
                  - 507 INSUFFICIENT_STORAGE
                  - 508 LOOP_DETECTED
                  - 509 BANDWIDTH_LIMIT_EXCEEDED
                  - 510 NOT_EXTENDED
                  - 511 NETWORK_AUTHENTICATION_REQUIRED
        "404":
          description: Not Found
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorMessage"
        "500":
          description: Internal Server Error
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorMessage"
...
  "/api/sections/{sectionsId}/students/{studentId}/means":
    delete:
      tags:
        - mean-controller
      summary: Delete a mean by sectionId and studentId
      operationId: deleteMeanOfstudentInSection
      parameters:
        - name: studentId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: sectionId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/Mean"
        "404":
          description: Not Found
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorMessage"
        "500":
          description: Internal Server Error
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorMessage"
...
components:
  schemas:
    ErrorMessage:
      type: object
      properties:
        statusCode:
          type: integer
          format: int32
        timestamp:
          type: string
          format: date-time
        message:
          type: string
        description:
          type: string
...
    Student:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
    Section:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        credits:
          type: integer
          format: int32
        level:
          type: integer
          format: int32
...
    Mean:
      type: object
      properties:
        id:
          type: integer
          format: int64
        mean:
          type: integer
          format: int32
        schoolYear:
          type: integer
          format: int32
        student:
          "$ref": "#/components/schemas/Student"
        section:
          "$ref": "#/components/schemas/Section"
...